+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * perf: Reduce RobotState size - reduce ethernet fragmentation
 * chore: Remove unused model library
 * fix: null pointer fix
+* feat: Added the async motion generator mode
 
 # 1.0.0
 * Added support to transmit the arm URDF to libfranka